extend getAuthorWorksData to all authors properties
