+++ server/controllers/entities/lib/entities.ts
@@ -13,12 +13,10 @@ import { emit } from '#lib/radio'
 import { assert_ } from '#lib/utils/assert_types'
 import { addEntityDocClaims, beforeEntityDocSave, setEntityDocLabels } from '#models/entity'
 import type { EntityImagePath, ImageHash } from '#server/types/image'
-import type { EntityUri, InvEntityDoc, EntityValue, PropertyUri, InvEntity, Isbn, InvClaimValue, SerializedEntity, WdEntityId, WdEntityUri, EntityType, InvEntityId } from '#types/entity'
+import type { EntityUri, InvEntityDoc, EntityValue, PropertyUri, InvEntity, Isbn, InvClaimValue, SerializedEntity, WdEntityId, WdEntityUri, EntityType } from '#types/entity'
 import { getInvEntityCanonicalUri } from './get_inv_entity_canonical_uri.js'
 import createPatch from './patches/create_patch.js'
-import { prefixifyInv } from './prefix.js'
 import { validateProperty } from './properties/validations.js'
-import { getEntityNormalizedTerms } from './terms_normalization.js'
 import type { DocumentViewResponse } from 'blue-cot/types/nano.js'
 
 const db = await dbFactory('entities')
@@ -198,20 +196,6 @@ export async function getCollectionEditions (workUri: EntityUri) {
   const editionsUris = await reverseClaims({ property: 'wdt:P195', value: workUri })
   return getEntitiesList(editionsUris)
 }
-
-export async function getAuthorWorksData (authorId: InvEntityId) {
-  const works = await getInvEntitiesByClaim('wdt:P50', `inv:${authorId}`, true, true)
-  // works = [
-  //   { labels: { fr: 'Matiere et Memoire'} },
-  //   { labels: { en: 'foo' } }
-  // ]
-  const labels = uniq(works.flatMap(getEntityNormalizedTerms))
-  const langs = uniq(works.flatMap(getLangs))
-  const worksUris = works.map(work => prefixifyInv(work._id))
-  return { authorId, labels, langs, worksUris }
-}
-
-const getLangs = work => Object.keys(work.labels)
 
 export async function getAuthorsFromWorksUris (workUris: EntityUri[]) {
   const works = await getEntitiesList(workUris)
