+++ server/controllers/tasks/lib/get_new_suggestions_or_automerge.ts
@@ -1,5 +1,4 @@
 import { map } from 'lodash-es'
-import { getAuthorWorksData } from '#controllers/entities/lib/entities'
 import { getEntitiesByUris } from '#controllers/entities/lib/get_entities_by_uris'
 import getOccurrencesFromEntities from '#controllers/entities/lib/get_occurrences_from_entities'
 import { getOccurrencesFromExternalSources } from '#controllers/entities/lib/get_occurrences_from_external_sources'
@@ -7,6 +6,7 @@ import { haveExactMatch } from '#controllers/entities/lib/labels_match'
 import { propertiesValuesConstraints as properties } from '#controllers/entities/lib/properties/properties_values_constraints'
 import { getEntityNormalizedTerms } from '#controllers/entities/lib/terms_normalization'
 import typeSearch from '#controllers/search/lib/type_search'
+import { getAuthorWorksData } from '#controllers/tasks/lib/get_author_works_data'
 import { isNonEmptyString } from '#lib/boolean_validations'
 import { forceArray, someMatch } from '#lib/utils/base'
 import config from '#server/config'
