+++ server/controllers/tasks/lib/merge_or_create_tasks.ts
@@ -1,9 +1,10 @@
 import { map, uniq } from 'lodash-es'
-import { getAuthorsFromWorksUris, getAuthorWorksData, getPublishersFromPublicationsUris } from '#controllers/entities/lib/entities'
+import { getAuthorsFromWorksUris, getPublishersFromPublicationsUris } from '#controllers/entities/lib/entities'
 import { getEntitiesList } from '#controllers/entities/lib/get_entities_list'
 import { haveExactMatch } from '#controllers/entities/lib/labels_match'
 import mergeEntities from '#controllers/entities/lib/merge_entities'
 import { getEntityNormalizedTerms } from '#controllers/entities/lib/terms_normalization'
+import { getAuthorWorksData } from '#controllers/tasks/lib/get_author_works_data'
 import { updateTasks, getExistingTasks, createTasksFromSuggestions, getTasksBySuspectUri } from '#controllers/tasks/lib/tasks'
 import { someMatch } from '#lib/utils/base'
 import type { SerializedEntity } from '#server/types/entity'
